JPAtoIFTTT.js
I am using this modified script on AWS to pass any command with any variable to the Maker Channel of IFTTT. You'll see by the Interaction Model, and the Sample Utterances that this script makes it possible to use nearly any command with your Echo to IFTTT. At this point, I am able to change my Hue lights to any basic color, or even use hex codes, all through the Echo.

Huge props and thanks to the actual creators of this code: mefogle/slack8181 and Monty617 as they are the true geniuses behind it all. Another big shout out to Matt Kruse whose Echottt project was what inspired me to get this to work.